Refactored: change admin categories folder naming
