+++ src/app/admin/categories/admin-categories.component.html
